remove all unused component and hook, remove binding folder
