+++ src/App.js
@@ -1,15 +1,10 @@
 import React, {useState, useEffect} from 'react'
-// import OldSection from './OldSection'
-import HomePage from './pages/HomePage'
-// import ImageContent from './pages/ImageContent'
-import ImageContent from './bindings/ImageContent'
+
+import ImageContent from './pages/ImageContent'
 
 function App() {
- 
   return (
     <div className="App">
-      {/* <OldSection /> */}
-      {/* <HomePage /> */}
       <ImageContent />
     </div>
   );
