--- src/components/Hero/index.jsx
@@ -1,24 +1,0 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-
-Hero.propTypes = {
-  name: PropTypes.string,
-}
-
-Hero.defaultProps = {
-  name: '',
-}
-
-
-function Hero(props) {
-  const { name } = props;
-  console.log('Here render: ', name);
-
-  return (
-    <div>
-      Hero name: {name}
-    </div>
-  )
-}
-
-export default React.memo(Hero);
