--- src/components/Home/HobbyList/HobbyList.css
@@ -1,11 +1,0 @@
-ul.hobby-list {
-  
-}
-
-ul.hobby-list > li {
-  transition: all .35s ease-in-out 0s;
-}
-
-ul.hobby-list > li.active {
-  color: deeppink;
-}
