--- src/components/MagicBox/MagicBox.scss
@@ -1,11 +1,0 @@
-.magic-box {
-  $box-size: 200px;
-
-  display: inline-block;
-  width: $box-size;
-  height: $box-size;
-  border-radius: 8px;
-
-  // border: solid 1px black;
-  transition: all 0.35s ease-in-out 0s;
-}
