--- src/components/MagicBox/index.jsx
@@ -1,16 +1,0 @@
-import React from 'react'
-import useMagicColor from '../../hooks/useMagicColor';
-import './MagicBox.scss'
-
-function MagicBox() {
-  const color = useMagicColor();
-
-  return (
-    <div
-      className="magic-box"
-      style={{ backgroundColor: color }}
-      ></div>
-  )
-}
-
-export default MagicBox;
